server/requirements.txt
sentence-transformers
numpy
openai
faiss-cpu
